change vehicle class and add abstraction
I applied the abstraction cconcept to this class and implement encapsulationts concepts for the attributes and methods. I also did some validation adjustments for the attributes and created the setter and getters needed. I erased the type attribute so I'll need to implement it as enumarators in the following change